ARM: at91/boards: use -EINVAL for invalid gpio
this will allow to use gpio_is_valid

Cc: Nicolas Ferre <nicolas.ferre@atmel.com>

+++ include/linux/platform_data/macb.h
@@ -11,7 +11,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 
 struct macb_platform_data {
 	u32		phy_mask;
-	u8		phy_irq_pin;	/* PHY IRQ */
+	int		phy_irq_pin;	/* PHY IRQ */
 	u8		is_rmii;	/* using RMII interface? */
 };
